backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
python-dotenv>=1.0.1
pydantic>=2.6.4
pyjwt>=2.10.1
python-multipart>=0.0.9
PyPDF2>=3.0.1
google-generativeai>=0.3.0
requests>=2.31.0
aiosqlite>=0.19.0
numpy>=1.26.0
scikit-learn>=1.3.0